Added Business Layer
Added Builder for entities
Added Service for entities
Added JUnit ServiceTest for services

+++ src/main/java/bootcamp/five/agency/newys/repository/ArticleRepository.java
@@ -12,6 +12,6 @@ public interface ArticleRepository extends JpaRepository<Article, Long> {
 
   List<Article> findByDateOfPublicationAfter(Date dateOfPublication);
 
-  List<Article> findByNumLikesGreaterThan(long numLikes);
+  List<Article> findByNumLikesGreaterThan(int numLikes);
 
 }
